Create Day of the year: writing and using your own functions.py
Objectives

Familiarize the student with:

    projecting and writing parameterized functions;
    utilizing the return statement;
    building a set of utility functions;
    utilizing the student's own functions.

Scenario

Your task is to write and test a function which takes three arguments (a year, a month, and a day of the month) and returns the corresponding day of the year, or returns None if any of the arguments is invalid.

Use the previously written and tested functions. Add some test cases to the code. This test is only a beginning.